Combine a4code2html tests into single file
